+++ src/new-components/password-field/password-field.stories.tsx
@@ -3,6 +3,7 @@ import { selectWithNone, PartialProps } from '../storybook-helper'
 import { PasswordField } from './'
 
 import type { BoxMaxWidth } from '../box'
+import { Stack } from '../stack'
 
 export default {
     title: 'Design system/PasswordField',
@@ -55,6 +56,15 @@ InteractivePropsStory.argTypes = {
         defaultValue:
             'Must be at least 100 characters long, and it should include each letter of the alphabet',
     },
+    message: {
+        control: { type: 'text' },
+        defaultValue: '',
+    },
+    tone: {
+        options: ['neutral', 'success', 'error', 'loading'],
+        control: { type: 'inline-radio' },
+        defaultValue: 'neutral',
+    },
     placeholder: {
         control: { type: 'text' },
         defaultValue: 'Type your password',
@@ -63,4 +73,37 @@ InteractivePropsStory.argTypes = {
         ['xsmall', 'small', 'medium', 'large', 'xlarge'],
         'small',
     ),
+}
+
+export function MessageToneStory() {
+    return (
+        <Stack space="xxlarge" dividers="secondary">
+            <PasswordField
+                label="Password confirmation"
+                message="Comparing to original password…"
+                tone="loading"
+                disabled
+                maxWidth="small"
+            />
+            <PasswordField
+                label="Password confirmation"
+                message="It does not match the original password"
+                tone="error"
+                maxWidth="small"
+            />
+            <PasswordField
+                label="Password confirmation"
+                message="Matches original password!"
+                tone="success"
+                maxWidth="small"
+            />
+            <PasswordField
+                label="Password confirmation"
+                message="Message with neutral tone (used as description, but still prefer the hint prop for that)"
+                hint="This is the primary description of the field, provided by the hint prop"
+                tone="neutral"
+                maxWidth="small"
+            />
+        </Stack>
+    )
 }