+++ src/new-components/text-field/text-field.stories.tsx
@@ -3,6 +3,7 @@ import { selectWithNone, PartialProps } from '../storybook-helper'
 import { TextField } from './'
 
 import type { BoxMaxWidth } from '../box'
+import { Stack } from '../stack'
 
 export default {
     title: 'Design system/TextField',
@@ -55,6 +56,15 @@ InteractivePropsStory.argTypes = {
         defaultValue:
             'We need your name for billing and shipping purposes. Make sure to enter it correctly.',
     },
+    message: {
+        control: { type: 'text' },
+        defaultValue: '',
+    },
+    tone: {
+        options: ['neutral', 'success', 'error', 'loading'],
+        control: { type: 'inline-radio' },
+        defaultValue: 'neutral',
+    },
     placeholder: {
         control: { type: 'text' },
         defaultValue: 'Enter your name as it appears in your ID',
@@ -63,4 +73,37 @@ InteractivePropsStory.argTypes = {
         ['xsmall', 'small', 'medium', 'large', 'xlarge'],
         'small',
     ),
+}
+
+export function MessageToneStory() {
+    return (
+        <Stack space="xxlarge" dividers="secondary">
+            <TextField
+                label="Verification code"
+                message="Verifying code…"
+                tone="loading"
+                disabled
+                maxWidth="small"
+            />
+            <TextField
+                label="Verification code"
+                message="Invalid code. Please, try again."
+                tone="error"
+                maxWidth="small"
+            />
+            <TextField
+                label="Verification code"
+                message="Code verification successful!"
+                tone="success"
+                maxWidth="small"
+            />
+            <TextField
+                label="Verification code"
+                message="Message with neutral tone (used as description, but still prefer the hint prop for that)"
+                hint="This is the primary description of the field, provided by the hint prop"
+                tone="neutral"
+                maxWidth="small"
+            />
+        </Stack>
+    )
 }